+++ README.md
@@ -15,14 +15,14 @@ For example, the output of one component can be used to define the input of anot
 
 ### Service descriptors
 
-Services should provide an implementation of [`ServiceDescriptor`](src/main/java/org/creek/api/platform/metadata/ServiceDescriptor.java)
+Services should provide an implementation of [`ServiceDescriptor`](src/main/java/org/creekservice/api/platform/metadata/ServiceDescriptor.java)
 to define key metadata needed by other components and the Creek system, including things like
 [inputs](#inputs), [internals](#internals), [outputs](#outputs), docker image name and any 
 environment variables to set during testing.
 
 ### Aggregate descriptors
 
-Each aggregate should provide an implementation of [`AggregateDescriptor`](src/main/java/org/creek/api/platform/metadata/AggregateDescriptor.java)
+Each aggregate should provide an implementation of [`AggregateDescriptor`](src/main/java/org/creekservice/api/platform/metadata/AggregateDescriptor.java)
 to define their public [inputs](#inputs) and [outputs](#outputs).
 
 ## Descriptors
@@ -33,19 +33,19 @@ Descriptors define, among other things, the components [inputs](#inputs), [inter
 
 The inputs of a component define any external resources the component consumes/reads.
 
-All input resources implement the [`ComponentInput`](src/main/java/org/creek/api/platform/metadata/ComponentInput.java)
+All input resources implement the [`ComponentInput`](src/main/java/org/creekservice/api/platform/metadata/ComponentInput.java)
 marker interface.
 
 ### Internals
 
 The internals of a component define any external resources the component uses internally to perform its function.
 
-All internal resources implement the [`ComponentInternal`](src/main/java/org/creek/api/platform/metadata/ComponentInternal.java)
+All internal resources implement the [`ComponentInternal`](src/main/java/org/creekservice/api/platform/metadata/ComponentInternal.java)
 marker interface.
 
 ### Outputs
 
 The outputs of a component define any external resources the component produces/writes.
 
-All output resources implement the [`ComponentOutput`](src/main/java/org/creek/api/platform/metadata/ComponentOutput.java)
+All output resources implement the [`ComponentOutput`](src/main/java/org/creekservice/api/platform/metadata/ComponentOutput.java)
 marker interface.
