+++ src/main/java/org/creekservice/api/platform/metadata/ResourceDescriptor.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 Creek Contributors (https://github.com/creek-service)
+ * Copyright 2021-2022 Creek Contributors (https://github.com/creek-service)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,16 +14,7 @@
  * limitations under the License.
  */
 
-package org.creek;
+package org.creekservice.api.platform.metadata;
 
-
-import org.creekservice.api.test.conformity.ConformityTester;
-import org.junit.jupiter.api.Test;
-
-class ModuleTest {
-
-    @Test
-    void shouldConform() {
-        ConformityTester.test(ModuleTest.class);
-    }
-}
+/** Marker interface of resource descriptors. */
+public interface ResourceDescriptor {}
